+++ lua/requests/views/autocmds.lua
@@ -6,12 +6,9 @@ local attach_quit_pre = function(state, bufnr)
         callback = function(p)
             ---@type number
             local win = state.win
-            state.win = nil
-            vim.g.requests_win = nil
             vim.api.nvim_win_close(win, true)
             if state.json_file then
                 vim.api.nvim_buf_delete(state.json_file, { force = true })
-                state.json_file = nil
             end
             for _, view in pairs(state.views) do
                 ---@type number
@@ -20,6 +17,7 @@ local attach_quit_pre = function(state, bufnr)
                     vim.api.nvim_buf_delete(buffer, { force = true })
                 end
             end
+            state.reset()
         end,
         buffer = bufnr,
     })