+++ plugin/init.lua
@@ -1,6 +1,14 @@
 local requests = require("requests")
 
-vim.api.nvim_create_user_command("Requests", function()
+local available_args = {
+    "empty",
+    "body",
+    "header",
+    "query",
+    "url",
+}
+
+vim.api.nvim_create_user_command("Requests", function(command)
     if vim.g.requests_win then
         vim.api.nvim_set_current_win(vim.g.requests_win)
         return
@@ -8,6 +16,18 @@ vim.api.nvim_create_user_command("Requests", function()
     if not vim.g.requests_setup then
         requests.setup()
     end
-    -- TODO: add autocomplete to go to specific View
-    requests.open("query")
-end, { desc = "Open Requests dashboard" })
+    local arg = command.fargs[1]
+    if arg == "empty" then
+        requests.open("query", true)
+        return
+    end
+    requests.open(arg)
+end, {
+    desc = "Open Requests dashboard",
+    nargs = "?",
+    complete = function(arg)
+        return vim.tbl_filter(function(value)
+            return vim.startswith(value, arg)
+        end, available_args)
+    end,
+})